pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["Anton Skopin <askopin@gmail.com>"]
readme = "README.md"
packages = [{include = "telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.4"
pytest = "^7.4.0"
ruff = "^0.0.277"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
